add drag atoms to residue selection menu
